+++ server/src/main/java/cn/ttplatform/wh/message/handler/RequestVoteResultMessageHandler.java
@@ -1,11 +1,11 @@
 package cn.ttplatform.wh.message.handler;
 
-import cn.ttplatform.wh.message.RequestVoteResultMessage;
-import cn.ttplatform.wh.constant.DistributableType;
-import cn.ttplatform.wh.Node;
 import cn.ttplatform.wh.GlobalContext;
+import cn.ttplatform.wh.Node;
+import cn.ttplatform.wh.constant.DistributableType;
 import cn.ttplatform.wh.group.Cluster;
 import cn.ttplatform.wh.group.Phase;
+import cn.ttplatform.wh.message.RequestVoteResultMessage;
 import cn.ttplatform.wh.role.Candidate;
 import cn.ttplatform.wh.support.AbstractDistributableHandler;
 import cn.ttplatform.wh.support.Distributable;
@@ -24,7 +24,7 @@ public class RequestVoteResultMessageHandler extends AbstractDistributableHandle
     }
 
     @Override
-    public int getHandlerType() {
+    public byte getHandlerType() {
         return DistributableType.REQUEST_VOTE_RESULT;
     }
